+++ tests/test_models.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import TYPE_CHECKING
 from unittest.mock import MagicMock
 from unittest.mock import call
 from unittest.mock import patch
 
 import pytest
+from django.db.models import DecimalField
 from pytest_django.asserts import assertQuerysetEqual
 
 from django_afip import exceptions
@@ -422,3 +424,33 @@ def test_receipt_entry_gt_total_discount() -> None:
 
     with pytest.raises(Exception, match=r"\bdiscount_less_than_total\b"):
         factories.ReceiptEntryFactory(quantity=1, unit_price=1, discount=2)
+
+
+@pytest.mark.django_db()
+def test_receipt_entry_has_quantity_decimal_field() -> None:
+    """
+    Test that ReceiptEntry quantity now is a DecimalField
+    """
+    field = models.ReceiptEntry._meta.get_field("quantity")
+
+    # Verifica que el campo ahora es de tipo DecimalField
+    assert isinstance(field, DecimalField)
+    assert field.max_digits == 15
+    assert field.decimal_places == 2
+
+
+@pytest.mark.django_db()
+def test_receipt_entry_manage_decimal_quantities() -> None:
+    """
+    Test that ReceiptEntry can manage a decimal quantity
+    """
+    factories.ReceiptEntryFactory(
+        quantity=Decimal("2"), unit_price=3.33
+    )
+    factories.ReceiptEntryFactory(
+        quantity=Decimal("5.23"), unit_price=3.33
+    )
+
+    assert models.ReceiptEntry.objects.all().count() == 2
+    assert models.ReceiptEntry.objects.first().quantity == Decimal("2")
+    assert models.ReceiptEntry.objects.last().quantity == Decimal("5.23")